refector: move previewChattingDto file
