--- src/components/Conversation/dto.ts
@@ -1,9 +1,0 @@
-interface User {
-  name: string;
-  message: string;
-}
-
-export interface PreviewChattingProps {
-  user: User;
-  onClick: () => void;
-}
